move isDomain to InternetAddressUtil
